Rewrite bump allocator and integrate it in main crate
