+++ Cargo.toml
@@ -12,9 +12,7 @@ rlibc = "1.0"
 spin = "0.4.5"
 volatile = "0.1.0"
 x86_64 = "0.1.2"
-
-[dependencies.hole_list_allocator]
-path = "libs/hole_list_allocator"
+linked_list_allocator = "0.4.2"
 
 [dependencies.lazy_static]
 features = ["spin_no_std"]
