--- libs/bump_allocator/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-authors = ["Philipp Oppermann <dev@phil-opp.com>"]
-name = "bump_allocator"
-version = "0.1.0"
-
-[dependencies]
-spin = "0.4.5"
