--- libs/hole_list_allocator/.gitignore
@@ -1,2 +1,0 @@
-# Generated by Cargo
-/target/
