--- libs/hole_list_allocator/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-authors = ["Philipp Oppermann <dev@phil-opp.com>"]
-name = "hole_list_allocator"
-version = "0.1.0"
-
-[dependencies]
-linked_list_allocator = { git = "https://github.com/phil-opp/linked-list-allocator.git"}
-spin = "0.4.5"
-
-[dependencies.lazy_static]
-version = "0.2.1"
-features = ["spin_no_std"]
